micro-ecommerce-api/requirements.txt
Django==3.0.7
django-environ==0.4.5
psycopg2==2.8.5
djangorestframework==3.11.0
django-polymorphic==2.1.2
django-rest-polymorphic==0.1.9
djangorestframework-simplejwt==4.3.0
django-filter==2.2.0
django-cors-headers==3.4.0
Pillow==7.1.2
django-autoslug==1.9.8
hashids==1.3.1
amqp==2.6.0
celery==4.4.3
kombu==4.6.9

